TIP-843: Rename array converters to enrichment
