+++ src/Akeneo/Pim/Enrichment/Component/Category/Connector/ArrayConverter/StandardToFlat/Category.php
@@ -3,7 +3,7 @@
 namespace Akeneo\Pim\Enrichment\Component\Category\Connector\ArrayConverter\StandardToFlat;
 
 use Pim\Component\Connector\ArrayConverter\ArrayConverterInterface;
-use Pim\Component\Connector\ArrayConverter\StandardToFlat\AbstractSimpleArrayConverter;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\StandardToFlat\AbstractSimpleArrayConverter;
 
 /**
  * Convert standard format to flat format for category
