+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/ArrayConverter/StandardToFlat/Product/ValueConverter/BooleanConverter.php
@@ -1,34 +1,34 @@
 <?php
 
-namespace Pim\Component\Connector\ArrayConverter\StandardToFlat\Product\ValueConverter;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\StandardToFlat\Product\ValueConverter;
 
 /**
- * Text array converter.
- * Convert a standard text array format to a flat one.
+ * Boolean array converter.
+ * Convert a standard boolean array format to a flat one.
  *
  * @author    Adrien Pétremann <adrien.petremann@akeneo.com>
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class TextConverter extends AbstractValueConverter implements ValueConverterInterface
+class BooleanConverter extends AbstractValueConverter implements ValueConverterInterface
 {
     /**
      * {@inheritdoc}
      *
-     * Convert a standard formatted text product value to a flat one.
+     * Convert a standard formatted boolean product value to a flat one.
      *
-     * Given a 'name' $attributeCode with this $data:
+     * Given a 'auto_lock' $property with this $data:
      * [
      *     [
      *         'locale' => 'de_DE',
      *         'scope'  => 'ecommerce',
-     *         'data'   => 'Wii U'
+     *         'data'   => false
      *     ],
      * ]
      *
      * It will return:
      * [
-     *     'name-de_DE-ecommerce' => 'Wii U',
+     *     'auto_lock-de_DE-ecommerce' => '0',
      * ]
      */
     public function convert($attributeCode, $data)
@@ -42,7 +42,7 @@ class TextConverter extends AbstractValueConverter implements ValueConverterInte
                 $value['scope']
             );
 
-            $convertedItem[$flatName] = (string) $value['data'];
+            $convertedItem[$flatName] = false === $value['data'] || null === $value['data'] ? '0' : '1';
         }
 
         return $convertedItem;
