+++ src/Akeneo/Pim/Structure/Component/ArrayConverter/StandardToFlat/FamilyVariant/FamilyVariant.php
@@ -3,7 +3,7 @@
 namespace Akeneo\Pim\Structure\Component\ArrayConverter\StandardToFlat\FamilyVariant;
 
 use Pim\Component\Connector\ArrayConverter\ArrayConverterInterface;
-use Pim\Component\Connector\ArrayConverter\StandardToFlat\AbstractSimpleArrayConverter;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\StandardToFlat\AbstractSimpleArrayConverter;
 
 /**
  * @author    Damien Carcel (damien.carcel@akeneo.com)
