+++ src/Akeneo/UserManagement/Component/Connector/ArrayConverter/StandardToFlat/User.php
@@ -3,7 +3,7 @@
 namespace Akeneo\UserManagement\Component\Connector\ArrayConverter\StandardToFlat;
 
 use Pim\Component\Connector\ArrayConverter\ArrayConverterInterface;
-use Pim\Component\Connector\ArrayConverter\StandardToFlat\AbstractSimpleArrayConverter;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ArrayConverter\StandardToFlat\AbstractSimpleArrayConverter;
 
 /**
  * Convert standard format to flat format for user
